Skybox Revamp; Grass Shader; Pulled Improvements from Template
Skybox shader was swapped for one that already supports a sun and moon disc, modified to better fit this game.

The grass shader was implemented to create an approximate BOTW style.

Improvements from the Squiggle Zone template were pulled in, specifically an improvement to the footstep sound system where by no sound will be played if the current detected object does not have a ground material attached